+++ templates/parts/loop/loop-products.php
@@ -13,7 +13,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.1.0
  */
 
 
@@ -37,23 +37,6 @@
 
 
 ?>
-
-<?php if ( $invoice_helper['dual_currency'] ) : ?>
-<div class="invoice-note invoice-note-exchange">
-	<?php
-
-	printf(
-		/* translators: 1: currency to exchange from, 2: currency to exchange to, 3: exchange rate, 4: invoice publish date. */
-		esc_html__( '%1$s to %2$s exchange rate of %3$f valid on %4$s (D/M/Y)', '_invoices' ),
-		$invoice_helper['currency_from'],
-		$invoice_helper['currency_to'],
-		$invoice_helper['exchange_rate'],
-		$invoice_helper['publish_date_display']
-	);
-
-	?>
-</div>
-<?php endif; ?>
 
 <table class="invoice-products">
 
@@ -100,7 +83,7 @@
 						?>
 						<td class="invoice-products-column-description">
 
-							<?php if ( $product = get_sub_field( 'product' ) ) : ?>
+							<?php if ( $product = get_post( get_sub_field( 'product' ) ) ) : ?>
 								<h3 class="product-title">
 									<?php echo esc_html( $product->post_title ); ?>
 								</h3>