+++ changelog.md
@@ -1,5 +1,28 @@
 # Invoices Changelog
 
+## 1.1.0
+
+* **Add**: Option to select default values for Invoice custom fields
+* **Add**: Option to select form field type for Invoice Clients
+* **Update**: Improving code organization
+* **Update**: Making currency exchange info a link to API data
+* **Update**: Localization
+* **Fix**: Print styles
+
+### Files changed:
+
+	changelog.md
+	functions.php
+	style.css
+	includes/class-advanced-custom-fields.php
+	includes/class-customize.php
+	includes/class-post-types.php
+	includes/class-taxonomies.php
+	languages/*.*
+	templates/parts/component/note-invoice.php
+	templates/parts/loop/loop-products.php
+
+
 ## 1.0.0
 
 * Initial release.